Use /metrics/snapshot instead of /master/state.json
This commit dramatically improves the response time and the stability
of the check_mesos.py check_mk check. On large Mesos cluster environments,
getting the state.json can take more than 10 sec at worst. --> Check timeouts.
Mesos is also providing a metrics endpoint (/metrics/snapshot) which is
much faster in response than the normal state endpoint (/master/state.json).
So the  metrics endpoint is now used to get cluster states.

Due to the fact that the metrics endpoint doesn't return the names of the
active frameworks, the check now ensures that the sum of active ones is bigger
than a given threshold (-F/--frameworks).
This replaces the behaviour of checking for a specific framework to be online.

Issue: INFRABRBA-4022